Crate icu_properties_data

Data for the icu_properties crate

This data was generated with CLDR version 47.0.0, ICU version icu4x/2025-05-01/77.x, and LSTM segmenter version v0.1.0.

Macros

impl_property_binary_alnum_v1

Implement DataProvider<PropertyBinaryAlnumV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_alphabetic_v1

Implement DataProvider<PropertyBinaryAlphabeticV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_ascii_hex_digit_v1

Implement DataProvider<PropertyBinaryAsciiHexDigitV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_basic_emoji_v1

Implement DataProvider<PropertyBinaryBasicEmojiV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_bidi_control_v1

Implement DataProvider<PropertyBinaryBidiControlV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_bidi_mirrored_v1

Implement DataProvider<PropertyBinaryBidiMirroredV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_blank_v1

Implement DataProvider<PropertyBinaryBlankV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_case_ignorable_v1

Implement DataProvider<PropertyBinaryCaseIgnorableV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_case_sensitive_v1

Implement DataProvider<PropertyBinaryCaseSensitiveV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_cased_v1

Implement DataProvider<PropertyBinaryCasedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_casefolded_v1

Implement DataProvider<PropertyBinaryChangesWhenCasefoldedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_casemapped_v1

Implement DataProvider<PropertyBinaryChangesWhenCasemappedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_lowercased_v1

Implement DataProvider<PropertyBinaryChangesWhenLowercasedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_nfkc_casefolded_v1

Implement DataProvider<PropertyBinaryChangesWhenNfkcCasefoldedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_titlecased_v1

Implement DataProvider<PropertyBinaryChangesWhenTitlecasedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_changes_when_uppercased_v1

Implement DataProvider<PropertyBinaryChangesWhenUppercasedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_dash_v1

Implement DataProvider<PropertyBinaryDashV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_default_ignorable_code_point_v1

Implement DataProvider<PropertyBinaryDefaultIgnorableCodePointV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_deprecated_v1

Implement DataProvider<PropertyBinaryDeprecatedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_diacritic_v1

Implement DataProvider<PropertyBinaryDiacriticV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_emoji_component_v1

Implement DataProvider<PropertyBinaryEmojiComponentV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_emoji_modifier_base_v1

Implement DataProvider<PropertyBinaryEmojiModifierBaseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_emoji_modifier_v1

Implement DataProvider<PropertyBinaryEmojiModifierV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_emoji_presentation_v1

Implement DataProvider<PropertyBinaryEmojiPresentationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_emoji_v1

Implement DataProvider<PropertyBinaryEmojiV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_extended_pictographic_v1

Implement DataProvider<PropertyBinaryExtendedPictographicV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_extender_v1

Implement DataProvider<PropertyBinaryExtenderV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_full_composition_exclusion_v1

Implement DataProvider<PropertyBinaryFullCompositionExclusionV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_graph_v1

Implement DataProvider<PropertyBinaryGraphV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_grapheme_base_v1

Implement DataProvider<PropertyBinaryGraphemeBaseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_grapheme_extend_v1

Implement DataProvider<PropertyBinaryGraphemeExtendV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_grapheme_link_v1

Implement DataProvider<PropertyBinaryGraphemeLinkV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_hex_digit_v1

Implement DataProvider<PropertyBinaryHexDigitV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_hyphen_v1

Implement DataProvider<PropertyBinaryHyphenV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_id_continue_v1

Implement DataProvider<PropertyBinaryIdContinueV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_id_start_v1

Implement DataProvider<PropertyBinaryIdStartV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_ideographic_v1

Implement DataProvider<PropertyBinaryIdeographicV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_ids_binary_operator_v1

Implement DataProvider<PropertyBinaryIdsBinaryOperatorV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_ids_trinary_operator_v1

Implement DataProvider<PropertyBinaryIdsTrinaryOperatorV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_join_control_v1

Implement DataProvider<PropertyBinaryJoinControlV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_logical_order_exception_v1

Implement DataProvider<PropertyBinaryLogicalOrderExceptionV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_lowercase_v1

Implement DataProvider<PropertyBinaryLowercaseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_math_v1

Implement DataProvider<PropertyBinaryMathV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_nfc_inert_v1

Implement DataProvider<PropertyBinaryNfcInertV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_nfd_inert_v1

Implement DataProvider<PropertyBinaryNfdInertV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_nfkc_inert_v1

Implement DataProvider<PropertyBinaryNfkcInertV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_nfkd_inert_v1

Implement DataProvider<PropertyBinaryNfkdInertV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_noncharacter_code_point_v1

Implement DataProvider<PropertyBinaryNoncharacterCodePointV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_pattern_syntax_v1

Implement DataProvider<PropertyBinaryPatternSyntaxV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_pattern_white_space_v1

Implement DataProvider<PropertyBinaryPatternWhiteSpaceV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_prepended_concatenation_mark_v1

Implement DataProvider<PropertyBinaryPrependedConcatenationMarkV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_print_v1

Implement DataProvider<PropertyBinaryPrintV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_quotation_mark_v1

Implement DataProvider<PropertyBinaryQuotationMarkV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_radical_v1

Implement DataProvider<PropertyBinaryRadicalV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_regional_indicator_v1

Implement DataProvider<PropertyBinaryRegionalIndicatorV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_segment_starter_v1

Implement DataProvider<PropertyBinarySegmentStarterV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_sentence_terminal_v1

Implement DataProvider<PropertyBinarySentenceTerminalV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_soft_dotted_v1

Implement DataProvider<PropertyBinarySoftDottedV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_terminal_punctuation_v1

Implement DataProvider<PropertyBinaryTerminalPunctuationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_unified_ideograph_v1

Implement DataProvider<PropertyBinaryUnifiedIdeographV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_uppercase_v1

Implement DataProvider<PropertyBinaryUppercaseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_variation_selector_v1

Implement DataProvider<PropertyBinaryVariationSelectorV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_white_space_v1

Implement DataProvider<PropertyBinaryWhiteSpaceV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_xdigit_v1

Implement DataProvider<PropertyBinaryXdigitV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_xid_continue_v1

Implement DataProvider<PropertyBinaryXidContinueV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_binary_xid_start_v1

Implement DataProvider<PropertyBinaryXidStartV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_bidi_class_v1

Implement DataProvider<PropertyEnumBidiClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_bidi_mirroring_glyph_v1

Implement DataProvider<PropertyEnumBidiMirroringGlyphV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_canonical_combining_class_v1

Implement DataProvider<PropertyEnumCanonicalCombiningClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_east_asian_width_v1

Implement DataProvider<PropertyEnumEastAsianWidthV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_general_category_v1

Implement DataProvider<PropertyEnumGeneralCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_grapheme_cluster_break_v1

Implement DataProvider<PropertyEnumGraphemeClusterBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_hangul_syllable_type_v1

Implement DataProvider<PropertyEnumHangulSyllableTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_indic_conjunct_break_v1

Implement DataProvider<PropertyEnumIndicConjunctBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_indic_syllabic_category_v1

Implement DataProvider<PropertyEnumIndicSyllabicCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_joining_type_v1

Implement DataProvider<PropertyEnumJoiningTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_line_break_v1

Implement DataProvider<PropertyEnumLineBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_script_v1

Implement DataProvider<PropertyEnumScriptV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_sentence_break_v1

Implement DataProvider<PropertyEnumSentenceBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_vertical_orientation_v1

Implement DataProvider<PropertyEnumVerticalOrientationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_enum_word_break_v1

Implement DataProvider<PropertyEnumWordBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_bidi_class_v1

Implement DataProvider<PropertyNameLongBidiClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_canonical_combining_class_v1

Implement DataProvider<PropertyNameLongCanonicalCombiningClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_east_asian_width_v1

Implement DataProvider<PropertyNameLongEastAsianWidthV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_general_category_v1

Implement DataProvider<PropertyNameLongGeneralCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_grapheme_cluster_break_v1

Implement DataProvider<PropertyNameLongGraphemeClusterBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_hangul_syllable_type_v1

Implement DataProvider<PropertyNameLongHangulSyllableTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_indic_syllabic_category_v1

Implement DataProvider<PropertyNameLongIndicSyllabicCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_joining_type_v1

Implement DataProvider<PropertyNameLongJoiningTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_line_break_v1

Implement DataProvider<PropertyNameLongLineBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_script_v1

Implement DataProvider<PropertyNameLongScriptV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_sentence_break_v1

Implement DataProvider<PropertyNameLongSentenceBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_vertical_orientation_v1

Implement DataProvider<PropertyNameLongVerticalOrientationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_long_word_break_v1

Implement DataProvider<PropertyNameLongWordBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_bidi_class_v1

Implement DataProvider<PropertyNameParseBidiClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_canonical_combining_class_v1

Implement DataProvider<PropertyNameParseCanonicalCombiningClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_east_asian_width_v1

Implement DataProvider<PropertyNameParseEastAsianWidthV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_general_category_mask_v1

Implement DataProvider<PropertyNameParseGeneralCategoryMaskV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_general_category_v1

Implement DataProvider<PropertyNameParseGeneralCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_grapheme_cluster_break_v1

Implement DataProvider<PropertyNameParseGraphemeClusterBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_hangul_syllable_type_v1

Implement DataProvider<PropertyNameParseHangulSyllableTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_indic_syllabic_category_v1

Implement DataProvider<PropertyNameParseIndicSyllabicCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_joining_type_v1

Implement DataProvider<PropertyNameParseJoiningTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_line_break_v1

Implement DataProvider<PropertyNameParseLineBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_script_v1

Implement DataProvider<PropertyNameParseScriptV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_sentence_break_v1

Implement DataProvider<PropertyNameParseSentenceBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_vertical_orientation_v1

Implement DataProvider<PropertyNameParseVerticalOrientationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_parse_word_break_v1

Implement DataProvider<PropertyNameParseWordBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_bidi_class_v1

Implement DataProvider<PropertyNameShortBidiClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_canonical_combining_class_v1

Implement DataProvider<PropertyNameShortCanonicalCombiningClassV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_east_asian_width_v1

Implement DataProvider<PropertyNameShortEastAsianWidthV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_general_category_v1

Implement DataProvider<PropertyNameShortGeneralCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_grapheme_cluster_break_v1

Implement DataProvider<PropertyNameShortGraphemeClusterBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_hangul_syllable_type_v1

Implement DataProvider<PropertyNameShortHangulSyllableTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_indic_syllabic_category_v1

Implement DataProvider<PropertyNameShortIndicSyllabicCategoryV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_joining_type_v1

Implement DataProvider<PropertyNameShortJoiningTypeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_line_break_v1

Implement DataProvider<PropertyNameShortLineBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_script_v1

Implement DataProvider<PropertyNameShortScriptV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_sentence_break_v1

Implement DataProvider<PropertyNameShortSentenceBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_vertical_orientation_v1

Implement DataProvider<PropertyNameShortVerticalOrientationV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_name_short_word_break_v1

Implement DataProvider<PropertyNameShortWordBreakV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_property_script_with_extensions_v1

Implement DataProvider<PropertyScriptWithExtensionsV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.