icu_properties_data

Macro impl_property_binary_pattern_syntax_v1

macro_rules! impl_property_binary_pattern_syntax_v1 {
    ($ provider : ty) => { ... };
    ($ provider : ty , ITER) => { ... };
    ($ provider : ty , DRY) => { ... };
    ($ provider : ty , DRY , ITER) => { ... };
}

Implement DataProvider<PropertyBinaryPatternSyntaxV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

Using this implementation will embed the following data in the binary’s data segment:

200B1 for the singleton data struct

This macro requires the following crates:

icu

icu_provider

zerovec

these numbers can be smaller in practice due to linker deduplication ↩