icu_properties_data

Macro impliterable_props_insc_v1

macro_rules! impliterable_props_insc_v1 {
    ($ provider : ty) => { ... };
}

Implement IterableDataProvider<IndicSyllabicCategoryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.