Crate icu_properties_data

Data for the icu_properties crate

This data was generated with CLDR version 45.0.0, ICU version icu4x/2024-05-16/75.x, and LSTM segmenter version v0.1.0.

Macros

impl_propnames_from_bc_v1

Implement DataProvider<BidiClassNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_ccc_v1

Implement DataProvider<CanonicalCombiningClassNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_ea_v1

Implement DataProvider<EastAsianWidthNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_gc_v1

Implement DataProvider<GeneralCategoryNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_gcb_v1

Implement DataProvider<GraphemeClusterBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_gcm_v1

Implement DataProvider<GeneralCategoryMaskNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_hst_v1

Implement DataProvider<HangulSyllableTypeNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_insc_v1

Implement DataProvider<IndicSyllabicCategoryNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_jt_v1

Implement DataProvider<JoiningTypeNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_lb_v1

Implement DataProvider<LineBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_sb_v1

Implement DataProvider<SentenceBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_sc_v1

Implement DataProvider<ScriptNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_from_wb_v1

Implement DataProvider<WordBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_bc_v1

Implement DataProvider<BidiClassValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_ea_v1

Implement DataProvider<EastAsianWidthValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_gc_v1

Implement DataProvider<GeneralCategoryValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_gcb_v1

Implement DataProvider<GraphemeClusterBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_hst_v1

Implement DataProvider<HangulSyllableTypeValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_insc_v1

Implement DataProvider<IndicSyllabicCategoryValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_jt_v1

Implement DataProvider<JoiningTypeValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_lb_v1

Implement DataProvider<LineBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_sb_v1

Implement DataProvider<SentenceBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_sc_v1

Implement DataProvider<ScriptValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_linear_wb_v1

Implement DataProvider<WordBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_long_sparse_ccc_v1

Implement DataProvider<CanonicalCombiningClassValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear4_sc_v1

Implement DataProvider<ScriptValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_bc_v1

Implement DataProvider<BidiClassValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_ea_v1

Implement DataProvider<EastAsianWidthValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_gc_v1

Implement DataProvider<GeneralCategoryValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_gcb_v1

Implement DataProvider<GraphemeClusterBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_hst_v1

Implement DataProvider<HangulSyllableTypeValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_insc_v1

Implement DataProvider<IndicSyllabicCategoryValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_jt_v1

Implement DataProvider<JoiningTypeValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_lb_v1

Implement DataProvider<LineBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_sb_v1

Implement DataProvider<SentenceBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_linear_wb_v1

Implement DataProvider<WordBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_propnames_to_short_sparse_ccc_v1

Implement DataProvider<CanonicalCombiningClassValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ahex_v1

Implement DataProvider<AsciiHexDigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_alnum_v1

Implement DataProvider<AlnumV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_alpha_v1

Implement DataProvider<AlphabeticV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_basic_emoji_v1

Implement DataProvider<BasicEmojiV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_bc_v1

Implement DataProvider<BidiClassV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_bidi_c_v1

Implement DataProvider<BidiControlV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_bidi_m_v1

Implement DataProvider<BidiMirroredV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_bidiauxiliaryprops_v1

Implement DataProvider<BidiAuxiliaryPropertiesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_blank_v1

Implement DataProvider<BlankV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cased_v1

Implement DataProvider<CasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ccc_v1

Implement DataProvider<CanonicalCombiningClassV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ci_v1

Implement DataProvider<CaseIgnorableV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_comp_ex_v1

Implement DataProvider<FullCompositionExclusionV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwcf_v1

Implement DataProvider<ChangesWhenCasefoldedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwcm_v1

Implement DataProvider<ChangesWhenCasemappedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwkcf_v1

Implement DataProvider<ChangesWhenNfkcCasefoldedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwl_v1

Implement DataProvider<ChangesWhenLowercasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwt_v1

Implement DataProvider<ChangesWhenTitlecasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_cwu_v1

Implement DataProvider<ChangesWhenUppercasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_dash_v1

Implement DataProvider<DashV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_dep_v1

Implement DataProvider<DeprecatedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_di_v1

Implement DataProvider<DefaultIgnorableCodePointV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_dia_v1

Implement DataProvider<DiacriticV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ea_v1

Implement DataProvider<EastAsianWidthV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ebase_v1

Implement DataProvider<EmojiModifierBaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ecomp_v1

Implement DataProvider<EmojiComponentV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_emod_v1

Implement DataProvider<EmojiModifierV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_emoji_v1

Implement DataProvider<EmojiV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_epres_v1

Implement DataProvider<EmojiPresentationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_exemplarchars_auxiliary_v1

Implement DataProvider<ExemplarCharactersAuxiliaryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_exemplarchars_index_v1

Implement DataProvider<ExemplarCharactersIndexV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_exemplarchars_main_v1

Implement DataProvider<ExemplarCharactersMainV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_exemplarchars_numbers_v1

Implement DataProvider<ExemplarCharactersNumbersV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_exemplarchars_punctuation_v1

Implement DataProvider<ExemplarCharactersPunctuationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ext_v1

Implement DataProvider<ExtenderV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_extpict_v1

Implement DataProvider<ExtendedPictographicV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_gc_v1

Implement DataProvider<GeneralCategoryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_gcb_v1

Implement DataProvider<GraphemeClusterBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_gr_base_v1

Implement DataProvider<GraphemeBaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_gr_ext_v1

Implement DataProvider<GraphemeExtendV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_gr_link_v1

Implement DataProvider<GraphemeLinkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_graph_v1

Implement DataProvider<GraphV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_hex_v1

Implement DataProvider<HexDigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_hst_v1

Implement DataProvider<HangulSyllableTypeV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_hyphen_v1

Implement DataProvider<HyphenV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_idc_v1

Implement DataProvider<IdContinueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ideo_v1

Implement DataProvider<IdeographicV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ids_v1

Implement DataProvider<IdStartV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_idsb_v1

Implement DataProvider<IdsBinaryOperatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_idst_v1

Implement DataProvider<IdsTrinaryOperatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_insc_v1

Implement DataProvider<IndicSyllabicCategoryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_join_c_v1

Implement DataProvider<JoinControlV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_jt_v1

Implement DataProvider<JoiningTypeV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_lb_v1

Implement DataProvider<LineBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_loe_v1

Implement DataProvider<LogicalOrderExceptionV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_lower_v1

Implement DataProvider<LowercaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_math_v1

Implement DataProvider<MathV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_nchar_v1

Implement DataProvider<NoncharacterCodePointV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_nfcinert_v1

Implement DataProvider<NfcInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_nfdinert_v1

Implement DataProvider<NfdInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_nfkcinert_v1

Implement DataProvider<NfkcInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_nfkdinert_v1

Implement DataProvider<NfkdInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_pat_syn_v1

Implement DataProvider<PatternSyntaxV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_pat_ws_v1

Implement DataProvider<PatternWhiteSpaceV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_pcm_v1

Implement DataProvider<PrependedConcatenationMarkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_print_v1

Implement DataProvider<PrintV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_qmark_v1

Implement DataProvider<QuotationMarkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_radical_v1

Implement DataProvider<RadicalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_ri_v1

Implement DataProvider<RegionalIndicatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_sb_v1

Implement DataProvider<SentenceBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_sc_v1

Implement DataProvider<ScriptV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_scx_v1

Implement DataProvider<ScriptWithExtensionsPropertyV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_sd_v1

Implement DataProvider<SoftDottedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_segstart_v1

Implement DataProvider<SegmentStarterV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_sensitive_v1

Implement DataProvider<CaseSensitiveV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_sterm_v1

Implement DataProvider<SentenceTerminalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_term_v1

Implement DataProvider<TerminalPunctuationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_uideo_v1

Implement DataProvider<UnifiedIdeographV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_upper_v1

Implement DataProvider<UppercaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_vs_v1

Implement DataProvider<VariationSelectorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_wb_v1

Implement DataProvider<WordBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_wspace_v1

Implement DataProvider<WhiteSpaceV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_xdigit_v1

Implement DataProvider<XdigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_xidc_v1

Implement DataProvider<XidContinueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_props_xids_v1

Implement DataProvider<XidStartV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impliterable_propnames_from_bc_v1

Implement IterableDataProvider<BidiClassNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_ccc_v1

Implement IterableDataProvider<CanonicalCombiningClassNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_ea_v1

Implement IterableDataProvider<EastAsianWidthNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_gc_v1

Implement IterableDataProvider<GeneralCategoryNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_gcb_v1

Implement IterableDataProvider<GraphemeClusterBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_gcm_v1

Implement IterableDataProvider<GeneralCategoryMaskNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_hst_v1

Implement IterableDataProvider<HangulSyllableTypeNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_insc_v1

Implement IterableDataProvider<IndicSyllabicCategoryNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_jt_v1

Implement IterableDataProvider<JoiningTypeNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_lb_v1

Implement IterableDataProvider<LineBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_sb_v1

Implement IterableDataProvider<SentenceBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_sc_v1

Implement IterableDataProvider<ScriptNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_from_wb_v1

Implement IterableDataProvider<WordBreakNameToValueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_bc_v1

Implement IterableDataProvider<BidiClassValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_ea_v1

Implement IterableDataProvider<EastAsianWidthValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_gc_v1

Implement IterableDataProvider<GeneralCategoryValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_gcb_v1

Implement IterableDataProvider<GraphemeClusterBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_hst_v1

Implement IterableDataProvider<HangulSyllableTypeValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_insc_v1

Implement IterableDataProvider<IndicSyllabicCategoryValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_jt_v1

Implement IterableDataProvider<JoiningTypeValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_lb_v1

Implement IterableDataProvider<LineBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_sb_v1

Implement IterableDataProvider<SentenceBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_sc_v1

Implement IterableDataProvider<ScriptValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_linear_wb_v1

Implement IterableDataProvider<WordBreakValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_long_sparse_ccc_v1

Implement IterableDataProvider<CanonicalCombiningClassValueToLongNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear4_sc_v1

Implement IterableDataProvider<ScriptValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_bc_v1

Implement IterableDataProvider<BidiClassValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_ea_v1

Implement IterableDataProvider<EastAsianWidthValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_gc_v1

Implement IterableDataProvider<GeneralCategoryValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_gcb_v1

Implement IterableDataProvider<GraphemeClusterBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_hst_v1

Implement IterableDataProvider<HangulSyllableTypeValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_insc_v1

Implement IterableDataProvider<IndicSyllabicCategoryValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_jt_v1

Implement IterableDataProvider<JoiningTypeValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_lb_v1

Implement IterableDataProvider<LineBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_sb_v1

Implement IterableDataProvider<SentenceBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_linear_wb_v1

Implement IterableDataProvider<WordBreakValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_propnames_to_short_sparse_ccc_v1

Implement IterableDataProvider<CanonicalCombiningClassValueToShortNameV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ahex_v1

Implement IterableDataProvider<AsciiHexDigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_alnum_v1

Implement IterableDataProvider<AlnumV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_alpha_v1

Implement IterableDataProvider<AlphabeticV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_basic_emoji_v1

Implement IterableDataProvider<BasicEmojiV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_bc_v1

Implement IterableDataProvider<BidiClassV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_bidi_c_v1

Implement IterableDataProvider<BidiControlV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_bidi_m_v1

Implement IterableDataProvider<BidiMirroredV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_bidiauxiliaryprops_v1

Implement IterableDataProvider<BidiAuxiliaryPropertiesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_blank_v1

Implement IterableDataProvider<BlankV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cased_v1

Implement IterableDataProvider<CasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ccc_v1

Implement IterableDataProvider<CanonicalCombiningClassV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ci_v1

Implement IterableDataProvider<CaseIgnorableV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_comp_ex_v1

Implement IterableDataProvider<FullCompositionExclusionV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwcf_v1

Implement IterableDataProvider<ChangesWhenCasefoldedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwcm_v1

Implement IterableDataProvider<ChangesWhenCasemappedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwkcf_v1

Implement IterableDataProvider<ChangesWhenNfkcCasefoldedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwl_v1

Implement IterableDataProvider<ChangesWhenLowercasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwt_v1

Implement IterableDataProvider<ChangesWhenTitlecasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_cwu_v1

Implement IterableDataProvider<ChangesWhenUppercasedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_dash_v1

Implement IterableDataProvider<DashV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_dep_v1

Implement IterableDataProvider<DeprecatedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_di_v1

Implement IterableDataProvider<DefaultIgnorableCodePointV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_dia_v1

Implement IterableDataProvider<DiacriticV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ea_v1

Implement IterableDataProvider<EastAsianWidthV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ebase_v1

Implement IterableDataProvider<EmojiModifierBaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ecomp_v1

Implement IterableDataProvider<EmojiComponentV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_emod_v1

Implement IterableDataProvider<EmojiModifierV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_emoji_v1

Implement IterableDataProvider<EmojiV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_epres_v1

Implement IterableDataProvider<EmojiPresentationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_exemplarchars_auxiliary_v1

Implement IterableDataProvider<ExemplarCharactersAuxiliaryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_exemplarchars_index_v1

Implement IterableDataProvider<ExemplarCharactersIndexV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_exemplarchars_main_v1

Implement IterableDataProvider<ExemplarCharactersMainV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_exemplarchars_numbers_v1

Implement IterableDataProvider<ExemplarCharactersNumbersV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_exemplarchars_punctuation_v1

Implement IterableDataProvider<ExemplarCharactersPunctuationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ext_v1

Implement IterableDataProvider<ExtenderV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_extpict_v1

Implement IterableDataProvider<ExtendedPictographicV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_gc_v1

Implement IterableDataProvider<GeneralCategoryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_gcb_v1

Implement IterableDataProvider<GraphemeClusterBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_gr_base_v1

Implement IterableDataProvider<GraphemeBaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_gr_ext_v1

Implement IterableDataProvider<GraphemeExtendV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_gr_link_v1

Implement IterableDataProvider<GraphemeLinkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_graph_v1

Implement IterableDataProvider<GraphV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_hex_v1

Implement IterableDataProvider<HexDigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_hst_v1

Implement IterableDataProvider<HangulSyllableTypeV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_hyphen_v1

Implement IterableDataProvider<HyphenV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_idc_v1

Implement IterableDataProvider<IdContinueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ideo_v1

Implement IterableDataProvider<IdeographicV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ids_v1

Implement IterableDataProvider<IdStartV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_idsb_v1

Implement IterableDataProvider<IdsBinaryOperatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_idst_v1

Implement IterableDataProvider<IdsTrinaryOperatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_insc_v1

Implement IterableDataProvider<IndicSyllabicCategoryV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_join_c_v1

Implement IterableDataProvider<JoinControlV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_jt_v1

Implement IterableDataProvider<JoiningTypeV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_lb_v1

Implement IterableDataProvider<LineBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_loe_v1

Implement IterableDataProvider<LogicalOrderExceptionV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_lower_v1

Implement IterableDataProvider<LowercaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_math_v1

Implement IterableDataProvider<MathV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_nchar_v1

Implement IterableDataProvider<NoncharacterCodePointV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_nfcinert_v1

Implement IterableDataProvider<NfcInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_nfdinert_v1

Implement IterableDataProvider<NfdInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_nfkcinert_v1

Implement IterableDataProvider<NfkcInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_nfkdinert_v1

Implement IterableDataProvider<NfkdInertV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_pat_syn_v1

Implement IterableDataProvider<PatternSyntaxV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_pat_ws_v1

Implement IterableDataProvider<PatternWhiteSpaceV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_pcm_v1

Implement IterableDataProvider<PrependedConcatenationMarkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_print_v1

Implement IterableDataProvider<PrintV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_qmark_v1

Implement IterableDataProvider<QuotationMarkV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_radical_v1

Implement IterableDataProvider<RadicalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_ri_v1

Implement IterableDataProvider<RegionalIndicatorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_sb_v1

Implement IterableDataProvider<SentenceBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_sc_v1

Implement IterableDataProvider<ScriptV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_scx_v1

Implement IterableDataProvider<ScriptWithExtensionsPropertyV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_sd_v1

Implement IterableDataProvider<SoftDottedV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_segstart_v1

Implement IterableDataProvider<SegmentStarterV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_sensitive_v1

Implement IterableDataProvider<CaseSensitiveV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_sterm_v1

Implement IterableDataProvider<SentenceTerminalV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_term_v1

Implement IterableDataProvider<TerminalPunctuationV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_uideo_v1

Implement IterableDataProvider<UnifiedIdeographV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_upper_v1

Implement IterableDataProvider<UppercaseV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_vs_v1

Implement IterableDataProvider<VariationSelectorV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_wb_v1

Implement IterableDataProvider<WordBreakV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_wspace_v1

Implement IterableDataProvider<WhiteSpaceV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_xdigit_v1

Implement IterableDataProvider<XdigitV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_xidc_v1

Implement IterableDataProvider<XidContinueV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_props_xids_v1

Implement IterableDataProvider<XidStartV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.